+++ sys/include/net/ipv6/hdr.h
@@ -61,7 +61,9 @@ extern "C" {
  * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
  * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  *
- * @see [RFC 8200, section 3](http://tools.ietf.org/html/rfc8200#section-3)
+ * @see <a href="http://tools.ietf.org/html/rfc2460#section-3">
+ *          RFC 2460, section 3
+ *      </a>
  */
 typedef struct __attribute__((packed)) {
     /**
@@ -266,7 +268,9 @@ static inline uint32_t ipv6_hdr_get_fl(const ipv6_hdr_t *hdr)
 /**
  * @brief   Calculates the Internet Checksum for the IPv6 Pseudo Header.
  *
- * @see [RFC 8200, section 8.1](https://tools.ietf.org/html/rfc8200#section-8.1)
+ * @see <a href="https://tools.ietf.org/html/rfc2460#section-8.1">
+ *          RFC 2460, section 8.1
+ *      </a>
  *
  * @param[in] sum       Preinialized value of the sum.
  * @param[in] prot_num  The @ref net_protnum you want to calculate the
@@ -284,7 +288,7 @@ static inline uint32_t ipv6_hdr_get_fl(const ipv6_hdr_t *hdr)
 static inline uint16_t ipv6_hdr_inet_csum(uint16_t sum, ipv6_hdr_t *hdr,
                                           uint8_t prot_num, uint16_t len)
 {
-    if (((uint32_t)sum + len + prot_num) > 0xffff) {
+    if ((sum + len + prot_num) > 0xffff) {
         /* increment by one for overflow to keep it as 1's complement sum */
         sum++;
     }
